requirements.txt
dotenv
requests
datetime
pandas
hyperliquid-python-sdk
numpy
psycopg2
websockets